requirements.txt
python-dotenv==1.0.0
python-telegram-bot~=13.1
redis==5.0.1